requirements.txt
fastapi==0.104.1
aiokafka==0.10.0
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
asyncpg==0.29.0
pydantic-settings==2.2.1
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
testcontainers==3.7.1
requests==2.31.0
asyncpg==0.29.0
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytest-asyncio==0.23.7